requirements.txt
# Core Dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1

# LangChain Dependencies
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.2

# Vector Database
qdrant-client==1.12.1

# ML/NLP
sentence-transformers==3.2.1

# File Processing
pypdf==5.1.0

# Security
bleach==6.2.0

# Production Server
gunicorn==21.2.0

# Development Dependencies (optional)
# python-multipart==0.0.6  # Required if handling file uploads
# pytest==8.2.0  # For testing
# pytest-cov==4.2.0  # For test coverage